requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
lightgbm>=4.0
scikit-learn>=1.3
tensorflow
pyyaml>=6.0
yfinance>=0.2
